+++ extension-impl/registry-sofa/src/main/java/com/alipay/sofa/rpc/registry/sofa/SofaRegistryConstants.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.rpc.common;
+package com.alipay.sofa.rpc.registry.sofa;
 
 /**
  * Created by zhanggeng on 2017/7/5.
