+++ extension-impl/registry-sofa/src/main/java/com/alipay/sofa/rpc/registry/sofa/SofaRegistryHelper.java
@@ -19,7 +19,6 @@ package com.alipay.sofa.rpc.registry.sofa;
 import com.alipay.sofa.rpc.client.ProviderInfo;
 import com.alipay.sofa.rpc.client.ProviderInfoAttrs;
 import com.alipay.sofa.rpc.client.ProviderStatus;
-import com.alipay.sofa.rpc.common.SofaRegistryConstants;
 import com.alipay.sofa.rpc.common.RemotingConstants;
 import com.alipay.sofa.rpc.common.RpcConstants;
 import com.alipay.sofa.rpc.common.SystemInfo;
@@ -52,21 +51,21 @@ import static com.alipay.sofa.rpc.client.ProviderInfoAttrs.ATTR_WARMUP_TIME;
 import static com.alipay.sofa.rpc.client.ProviderInfoAttrs.ATTR_WARMUP_WEIGHT;
 import static com.alipay.sofa.rpc.client.ProviderInfoAttrs.ATTR_WARM_UP_END_TIME;
 import static com.alipay.sofa.rpc.client.ProviderInfoAttrs.ATTR_WEIGHT;
-import static com.alipay.sofa.rpc.common.SofaRegistryConstants.APP_NAME;
-import static com.alipay.sofa.rpc.common.SofaRegistryConstants.HOST_MACHINE_KEY;
-import static com.alipay.sofa.rpc.common.SofaRegistryConstants.KEY_TIMEOUT;
-import static com.alipay.sofa.rpc.common.SofaRegistryConstants.RPC_REMOTING_PROTOCOL;
-import static com.alipay.sofa.rpc.common.SofaRegistryConstants.RPC_SERVICE_VERSION;
-import static com.alipay.sofa.rpc.common.SofaRegistryConstants.SERIALIZE_TYPE_KEY;
-import static com.alipay.sofa.rpc.common.SofaRegistryConstants.SOFA4_RPC_SERVICE_VERSION;
-import static com.alipay.sofa.rpc.common.SofaRegistryConstants.TIMEOUT;
-import static com.alipay.sofa.rpc.common.SofaRegistryConstants.WEIGHT_KEY;
 import static com.alipay.sofa.rpc.common.RpcConstants.PROTOCOL_TYPE_BOLT;
 import static com.alipay.sofa.rpc.common.RpcConstants.PROTOCOL_TYPE_TR;
 import static com.alipay.sofa.rpc.common.RpcConstants.SERIALIZE_HESSIAN;
 import static com.alipay.sofa.rpc.common.RpcConstants.SERIALIZE_HESSIAN2;
 import static com.alipay.sofa.rpc.common.RpcConstants.SERIALIZE_JAVA;
 import static com.alipay.sofa.rpc.common.RpcConstants.SERIALIZE_PROTOBUF;
+import static com.alipay.sofa.rpc.registry.sofa.SofaRegistryConstants.APP_NAME;
+import static com.alipay.sofa.rpc.registry.sofa.SofaRegistryConstants.HOST_MACHINE_KEY;
+import static com.alipay.sofa.rpc.registry.sofa.SofaRegistryConstants.KEY_TIMEOUT;
+import static com.alipay.sofa.rpc.registry.sofa.SofaRegistryConstants.RPC_REMOTING_PROTOCOL;
+import static com.alipay.sofa.rpc.registry.sofa.SofaRegistryConstants.RPC_SERVICE_VERSION;
+import static com.alipay.sofa.rpc.registry.sofa.SofaRegistryConstants.SERIALIZE_TYPE_KEY;
+import static com.alipay.sofa.rpc.registry.sofa.SofaRegistryConstants.SOFA4_RPC_SERVICE_VERSION;
+import static com.alipay.sofa.rpc.registry.sofa.SofaRegistryConstants.TIMEOUT;
+import static com.alipay.sofa.rpc.registry.sofa.SofaRegistryConstants.WEIGHT_KEY;
 
 /**
  * Created by zhanggeng on 2017/7/5.
@@ -352,9 +351,9 @@ public class SofaRegistryHelper {
         }
         // TODO SOFAVERSION v=4.0
         // timeout 
-        String timeoutStr = getValue(parameters, ATTR_TIMEOUT, SofaRegistryConstants.TIMEOUT);
+        String timeoutStr = getValue(parameters, ATTR_TIMEOUT, TIMEOUT);
         if (timeoutStr != null) {
-            removeOldKeys(parameters, ATTR_TIMEOUT, SofaRegistryConstants.TIMEOUT);
+            removeOldKeys(parameters, ATTR_TIMEOUT, TIMEOUT);
             try {// 加入动态
                 providerInfo.setDynamicAttr(ATTR_TIMEOUT, Integer.parseInt(timeoutStr));
             } catch (Exception e) {
@@ -363,9 +362,9 @@ public class SofaRegistryHelper {
         }
         // serializeType 使用字符传递
         String serializationStr = getValue(parameters, ATTR_SERIALIZATION,
-            SofaRegistryConstants.SERIALIZE_TYPE_KEY);
+            SERIALIZE_TYPE_KEY);
         if (serializationStr != null) {
-            removeOldKeys(parameters, ATTR_SERIALIZATION, SofaRegistryConstants.SERIALIZE_TYPE_KEY);
+            removeOldKeys(parameters, ATTR_SERIALIZATION, SERIALIZE_TYPE_KEY);
             // 1 -> hessian   2->java   4->hessian2  11->protobuf
             if ((RemotingConstants.SERIALIZE_CODE_HESSIAN + "").equals(serializationStr)) {
                 serializationStr = SERIALIZE_HESSIAN;
@@ -379,10 +378,10 @@ public class SofaRegistryHelper {
             providerInfo.setSerializationType(serializationStr);
         }
         // appName
-        String appNameStr = getValue(parameters, ATTR_APP_NAME, SofaRegistryConstants.APP_NAME,
+        String appNameStr = getValue(parameters, ATTR_APP_NAME, APP_NAME,
             SofaRegistryConstants.SELF_APP_NAME);
         if (appNameStr != null) {
-            removeOldKeys(parameters, SofaRegistryConstants.APP_NAME, SofaRegistryConstants.SELF_APP_NAME);
+            removeOldKeys(parameters, APP_NAME, SofaRegistryConstants.SELF_APP_NAME);
             providerInfo.setStaticAttr(ATTR_APP_NAME, appNameStr);
         }
         // connections
@@ -392,9 +391,9 @@ public class SofaRegistryHelper {
             providerInfo.setStaticAttr(ATTR_CONNECTIONS, connections);
         }
         // weight
-        String weightStr = getValue(parameters, ATTR_WEIGHT, SofaRegistryConstants.WEIGHT_KEY);
+        String weightStr = getValue(parameters, ATTR_WEIGHT, WEIGHT_KEY);
         if (weightStr != null) {
-            removeOldKeys(parameters, ATTR_WEIGHT, SofaRegistryConstants.WEIGHT_KEY);
+            removeOldKeys(parameters, ATTR_WEIGHT, WEIGHT_KEY);
             try {
                 int weight = Integer.parseInt(weightStr);
                 providerInfo.setWeight(weight);
@@ -505,10 +504,10 @@ public class SofaRegistryHelper {
                 }
             }
             // timeout特殊处理
-            String timeout = getValue(tmp, ATTR_TIMEOUT, SofaRegistryConstants.KEY_TIMEOUT,
-                SofaRegistryConstants.TIMEOUT);
+            String timeout = getValue(tmp, ATTR_TIMEOUT, KEY_TIMEOUT,
+                TIMEOUT);
             if (timeout != null) {
-                removeOldKeys(tmp, ATTR_TIMEOUT, SofaRegistryConstants.KEY_TIMEOUT, SofaRegistryConstants.TIMEOUT);
+                removeOldKeys(tmp, ATTR_TIMEOUT, KEY_TIMEOUT, TIMEOUT);
                 try {
                     methodParameters.put("." + method + "." + ATTR_TIMEOUT,
                         Integer.parseInt(timeout));
